featureSearchBar connected
