+++ client/src/pages/SearchResults/styledSeachResults.js
@@ -1,0 +1,1 @@
+import styled from "styled-components";
